Move query to separate file
